# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DB API — Multitenant REST API",
        "description": "Laravel 13 API with DDD/CQRS architecture and table-per-tenant multi-tenancy.\n\n## Authentication\nAll write endpoints require a Bearer token issued by Laravel Passport.\nObtain one via `POST /oauth/token` (standard Passport password grant).\n\n## Multitenancy\nEvery request must target a tenant via the URL `/tenant_id/v1/...`.\nTenant identifiers are 3-63 lowercase alphanumeric characters (hyphens/underscores allowed).\n\n## Module gating\nSome endpoints require the tenant to have a specific module enabled.\nIf a module is disabled, the API returns **403** with the message\n`\"Module 'X' is not enabled for this tenant\"`.\n\n## Standard response envelope\nAll endpoints return a consistent JSON envelope:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"message\": \"Operation completed successfully\"\n}\n```\nErrors replace `data` with `errors` (validation details).",
        "contact": {
            "email": "dev@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development"
        },
        {
            "url": "https://api.example.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/{tenant}/{version}/posts": {
            "get": {
                "tags": [
                    "Blogging - Posts"
                ],
                "summary": "List posts (filterable, paginated)",
                "operationId": "432df7e5b879677f91307cc3b1c7ee55",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "language_code",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "en"
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of posts"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogging - Posts"
                ],
                "summary": "Crear un post de blog",
                "operationId": "99e7572cf49719ec14a7d8d69ec9a70d",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "acme"
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "v1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title",
                                    "content",
                                    "language"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Mi primer post con IA",
                                        "maxLength": 255
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "La inteligencia artificial es increíble..."
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "es",
                                        "maxLength": 5
                                    },
                                    "seo_title": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 60
                                    },
                                    "seo_description": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 160
                                    },
                                    "seo_keywords": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "canonical_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "og_title": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 95
                                    },
                                    "og_description": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 200
                                    },
                                    "og_image": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "tag_names": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "php",
                                            "laravel",
                                            "ia"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/{tenant}/{version}/forms/{id}": {
            "get": {
                "tags": [
                    "Forms"
                ],
                "summary": "Get form definition",
                "description": "Retrieve form schema including field definitions. Authenticated only.",
                "operationId": "5460ad11acc5a6b771f833948165d8d8",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Form ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Form not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "description": "Standard success response envelope",
                "properties": {
                    "success": {
                        "description": "OpenAPI specification for DB API.\n\nCentral definition file — every annotation here is merged with per-controller\nannotations scanned by l5-swagger (scan paths: app/, src/).",
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "example": null,
                        "nullable": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Standard error response envelope",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "FormResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "key": {
                        "type": "string"
                    },
                    "recipient_email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "label": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "text",
                                        "email",
                                        "tel",
                                        "textarea",
                                        "select",
                                        "checkbox"
                                    ]
                                },
                                "required": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Passport",
                "description": "Laravel Passport token. Obtain via POST /oauth/token (password grant)."
            }
        }
    },
...
}